no commit message

+++ quanmin_admin/ybg_company/src/main/java/com/ybg/oa/department/controller/DepartmentController.java
@@ -21,6 +21,7 @@ import com.ybg.base.util.Json;
 import com.ybg.base.util.Page;
 import com.ybg.oa.department.domain.DepartmentVO;
 import com.ybg.oa.department.service.DepartmentService;
+import com.ybg.oa.constant.CompanyConstant;
 import com.ybg.oa.department.domain.DepartmentDO;
 import com.ybg.oa.department.qvo.DepartmentQuery;
 import springfox.documentation.annotations.ApiIgnore;
@@ -28,7 +29,7 @@ import springfox.documentation.annotations.ApiIgnore;
 /** @author Deament
  * @email 591518884@qq.com
  * @date 2017-07-15 */
-@Api(tags="Department管理")
+@Api(tags = "Department管理")
 @Controller
 @RequestMapping("/oa/department_do/")
 public class DepartmentController {
@@ -43,25 +44,16 @@ public class DepartmentController {
 	}
 	
 	@ApiOperation(value = "Department分页列表", notes = "JSON ", produces = MediaType.APPLICATION_JSON_VALUE)
-	//@ApiImplicitParams({ @ApiImplicitParam(name = "pageNow", value = "当前页数", required = true, dataType = "Integer"), @ApiImplicitParam(name = "qvo", value = "查询页数", required = false, dataType = "DepartmentQvo") })
+	// @ApiImplicitParams({ @ApiImplicitParam(name = "pageNow", value = "当前页数", required = true, dataType = "Integer"), @ApiImplicitParam(name = "qvo", value = "查询页数", required = false, dataType = "DepartmentQvo") })
 	@ResponseBody
 	@RequestMapping(value = { "list.do" }, method = { RequestMethod.GET, RequestMethod.POST })
 	public Page list(@ModelAttribute DepartmentQuery qvo, @ModelAttribute Page page, ModelMap map) throws Exception {
 		qvo.setBlurred(true);
-//		Page page = new Page();
-//		page.setCurPage(pageNow);
+		// Page page = new Page();
+		// page.setCurPage(pageNow);
 		page = departmentService.list(page, qvo);
 		page.init();
 		return page;
-	}
-	
-	/** 新增初始化
-	 * 
-	 * @throws Exception **/
-	@ApiOperation(value = "添加Department页面", notes = "", produces = MediaType.TEXT_HTML_VALUE)
-	@RequestMapping(value = { "toadd.do" }, method = { RequestMethod.GET, RequestMethod.POST })
-	public String toadd(@ApiIgnore ModelMap map) throws Exception {
-		return "/system/department/toadd";
 	}
 	
 	@ApiOperation(value = "更新Department", notes = "", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -73,10 +65,8 @@ public class DepartmentController {
 		try {
 			BaseMap<String, Object> updatemap = new BaseMap<String, Object>();
 			updatemap.put("name", department.getName());
-			updatemap.put("companyid", department.getCompanyid());
-			updatemap.put("gmt_create", department.getGmtCreate());
+			// updatemap.put("companyid", department.getCompanyid());
 			updatemap.put("parentid", department.getParentid());
-			updatemap.put("gmt_modified", department.getGmtModified());
 			updatemap.put("companyname", department.getCompanyname());
 			BaseMap<String, Object> wheremap = new BaseMap<String, Object>();
 			wheremap.put("id", department.getId());
@@ -100,6 +90,10 @@ public class DepartmentController {
 		try {
 			String[] ids = ids2.split(",");
 			for (String id : ids) {
+				if (id.equals(CompanyConstant.DEFAULT_DEPARTMENT_ID)) {
+					// 不删除默认设置
+					continue;
+				}
 				BaseMap<String, Object> wheremap = new BaseMap<String, Object>();
 				wheremap.put("id", id);
 				departmentService.remove(wheremap);
@@ -128,13 +122,5 @@ public class DepartmentController {
 		}
 		j.setMsg("操作成功");
 		return j;
-	}
-	
-	@ApiOperation(value = "更新department页面初始化", notes = " ", produces = MediaType.TEXT_HTML_VALUE)
-	@ApiImplicitParam(name = "id", value = "department的ID", required = true, dataType = "java.lang.String")
-	@RequestMapping(value = { "toupdate.do" }, method = { RequestMethod.GET, RequestMethod.POST })
-	public String toupdate(@RequestParam(name = "id", required = true) String id, ModelMap map) throws Exception {
-		map.put("department", departmentService.get(id));
-		return "/system/department/edit";
 	}
 }
